pyproject.toml
[tool.poetry]
name = "streamlit-sample-app"
version = "0.1.0"
description = ""
authors = ["INO <katsuki104@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.33.0"
openai = "=1.23.6"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
html2text = "^2024.2.26"
chromadb = "=0.4.10"
tiktoken = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
